Updated models, added comments and migrations

+++ Controllers/Admin/AuthorController.php
@@ -3,10 +3,21 @@
 namespace ChickenTikkaMasala\LaraCms\Controllers\Admin;
 
 use ChickenTikkaMasala\LaraCms\Controllers\Controller;
+use ChickenTikkaMasala\LaraCms\Models\Author;
+use ChickenTikkaMasala\LaraCms\Models\Site;
 use Illuminate\Http\Request;
 
+/**
+ * Class AuthorController
+ * @package ChickenTikkaMasala\LaraCms\Controllers\Admin
+ */
 class AuthorController extends Controller
 {
+    /**
+     * @param Request $request
+     * @param Site $site
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function index(Request $request, Site $site)
     {
         $authors = $site->authors();
@@ -23,13 +34,26 @@ class AuthorController extends Controller
         return response()->json($authors->paginate(10));
     }
 
+    /**
+     * @param Site $site
+     * @param Author $author
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function show(Site $site, Author $author)
     {
         return response()->json($author);
     }
 
+    /**
+     * @param Request $request
+     * @param Site $site
+     * @param Author $author
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function update(Request $request, Site $site, Author $author)
     {
+        $this->authorize('edit', Author::class);
+
         $this->validate($request, $this->rules);
 
         $author->update($request->all());
@@ -37,8 +61,15 @@ class AuthorController extends Controller
         return response()->json($author);
     }
 
+    /**
+     * @param Request $request
+     * @param Site $site
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function store(Request $request, Site $site)
     {
+        $this->authorize('create', Author::class);
+
         $this->validate($request, $this->rules);
 
         $author = $site->authors()->create($request->all());
@@ -46,8 +77,16 @@ class AuthorController extends Controller
         return response()->json($author);
     }
 
+    /**
+     * @param Request $request
+     * @param Site $site
+     * @param Author $author
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function destroy(Request $request, Site $site, Author $author)
     {
+        $this->authorize('destroy', Author::class);
+
         $author->destroy();
 
         return response()->json([
